Replace DiskFileSystem with RealFileSystem and clean up imports across the project
