--- crates/moss-fs/src/adapters.rs
@@ -1,1 +1,0 @@
-pub mod disk;
