+++ crates/moss-theme/src/theme_service.rs
@@ -1,6 +1,6 @@
 use anyhow::{anyhow, Result};
 use moss_app::service::prelude::AppService;
-use moss_fs::ports::FileSystem;
+use moss_fs::FileSystem;
 use std::{collections::HashMap, path::PathBuf, sync::Arc};
 use tokio::sync::OnceCell;
